Devices: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@62920 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/testcase/tstIntNetR0.cpp
@@ -194,7 +194,7 @@ static int tstIntNetSendBuf(PINTNETRINGBUF pRingBuf, INTNETIFHANDLE hIf,
                             PSUPDRVSESSION pSession, void const *pvBuf, size_t cbBuf)
 {
     INTNETSG Sg;
-    IntNetSgInitTemp(&Sg, (void *)pvBuf, cbBuf);
+    IntNetSgInitTemp(&Sg, (void *)pvBuf, (uint32_t)cbBuf);
     int rc = intnetR0RingWriteFrame(pRingBuf, &Sg, NULL);
     if (RT_SUCCESS(rc))
         rc = IntNetR0IfSend(hIf, pSession);
